refactor: entities and features
