+++ README.md
@@ -189,6 +189,9 @@ Clear.args = {
 - [articleEditForm](/src/features/articleEditForm)
 - [articleRating](/src/features/articleRating)
 - [articleRecommendationsList](/src/features/articleRecommendationsList)
+- [ArticleSortSelector](/src/features/ArticleSortSelector)
+- [ArticleTypeTabs](/src/features/ArticleTypeTabs)
+- [ArticleViewSelector](/src/features/ArticleViewSelector)
 - [AuthByUsername](/src/features/AuthByUsername)
 - [avatarDropdown](/src/features/avatarDropdown)
 - [editableProfileCard](/src/features/editableProfileCard)
