+++ src/entities/Article/README.md
@@ -11,12 +11,6 @@
 
 `ArticleList` -  Компонент со списком статей
 
-`ArticleViewSelector` - Компонент переключатель отображения списка статьей (плитка, список)
-
-`ArticleSortSelector` - Компонент с выбором сортировки списка статьей
-
-`ArticleTypeTabs` - Компонент с выбором типа статьи
-
 - types
 
 `Article` - Тип, описывающий статью
