+++ src/entities/Article/index.ts
@@ -7,6 +7,4 @@ export type { Article } from './model/types/article';
 export type { ArticleDetailsSchema } from './model/types/articleDetailsSchema';
 
 export { ArticleList } from './ui/ArticleList/ArticleList';
-export { ArticleViewSelector } from './ui/ArticleViewSelector/ArticleViewSelector';
-export { ArticleSortSelector } from './ui/ArticleSortSelector/ArticleSortSelector';
 export { getArticleDetailsData } from './model/selectors/articleDetails';
