+++ src/features/ArticleSortSelector/index.ts
@@ -1,0 +1,1 @@
+export { ArticleSortSelector } from './ui/ArticleSortSelector';
