+++ src/features/ArticleSortSelector/ui/ArticleSortSelector.tsx
@@ -1,9 +1,9 @@
 import { useTranslation } from 'react-i18next';
 import { memo, useMemo } from 'react';
-import { classNames } from '../../../../shared/lib/classNames/classNames';
-import { Select, SelectOption } from '../../../../shared/ui/Select/Select';
-import { ArticleSortField } from '../../../../entities/Article/model/consts/articleConsts';
-import { SortOrder } from '../../../../shared/types';
+import { classNames } from '../../../shared/lib/classNames/classNames';
+import { Select, SelectOption } from '../../../shared/ui/Select/Select';
+import { ArticleSortField } from '../../../entities/Article/model/consts/articleConsts';
+import { SortOrder } from '../../../shared/types';
 import cls from './ArticleSortSelector.module.scss';
 
 interface ArticleSortSelectorProps {
