+++ src/features/ArticleTypeTabs/README.md
@@ -1,0 +1,1 @@
+## Фича для выбора типа статьи
