+++ src/features/ArticleTypeTabs/index.ts
@@ -1,0 +1,1 @@
+export { ArticleTypeTabs } from './ui/ArticleTypeTabs';
