+++ src/features/ArticleTypeTabs/ui/ArticleTypeTabs.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next';
 import { memo, useCallback, useMemo } from 'react';
-import { classNames } from '../../../../shared/lib/classNames/classNames';
-import { TabItem, Tabs } from '../../../../shared/ui/Tabs/Tabs';
-import { ArticleType } from '../../../../entities/Article';
+import { classNames } from '../../../shared/lib/classNames/classNames';
+import { TabItem, Tabs } from '../../../shared/ui/Tabs/Tabs';
+import { ArticleType } from '../../../entities/Article';
 
 interface ArticleTypeTabsProps {
     className?: string;
