+++ src/features/ArticleViewSelector/README.md
@@ -1,0 +1,1 @@
+## Фича-переключатель для отображения списка статьей (плитка, список)
