+++ src/features/ArticleViewSelector/index.ts
@@ -1,0 +1,1 @@
+export { ArticleViewSelector } from './ui/ArticleViewSelector';
