+++ src/features/ArticleViewSelector/ui/ArticleViewSelector.tsx
@@ -1,11 +1,11 @@
 import { memo } from 'react';
 import TiledIcon from '../../../../shared/assets/icons/tiled-24-24.svg';
-import { Icon } from '../../../../shared/ui/Icon/Icon';
-import { Button, ButtonTheme } from '../../../../shared/ui/Button/Button';
+import { Icon } from '../../../shared/ui/Icon/Icon';
+import { Button, ButtonTheme } from '../../../shared/ui/Button/Button';
 import ListIcon from '../../../../shared/assets/icons/list-24-24.svg';
-import { classNames } from '../../../../shared/lib/classNames/classNames';
+import { classNames } from '../../../shared/lib/classNames/classNames';
 import cls from './ArticleViewSelector.module.scss';
-import { ArticleView } from '../../model/consts/articleConsts';
+import { ArticleView } from '../../../entities/Article/model/consts/articleConsts';
 
 interface ArticleViewSelectorProps {
     className?: string;
