+++ src/pages/ArticlesPage/ui/ArticlesPageFilters/ArticlesPageFilters.tsx
@@ -1,17 +1,13 @@
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
-import { ArticleTypeTabs } from '../../../../entities/Article/ui/ArticleTypeTabs/ArticleTypeTabs';
+import { ArticleTypeTabs } from '../../../../features/ArticleTypeTabs';
 import { useDebounce } from '../../../../shared/lib/hooks/useDebounce/useDebounce';
 import { fetchArticlesList } from '../../../../pages/ArticlesPage/model/services/fetchArticlesList/fetchArticlesList';
-import { ArticleSortSelector } from '../../../../entities/Article/ui/ArticleSortSelector/ArticleSortSelector';
 import { SortOrder } from '../../../../shared/types';
 import { Card } from '../../../../shared/ui/Card/Card';
 import { Input } from '../../../../shared/ui/Input/Input';
-import { Select } from '../../../../shared/ui/Select/Select';
-import {
-    ArticleSortField, ArticleType, ArticleView, ArticleViewSelector,
-} from '../../../../entities/Article';
+import { ArticleSortField, ArticleType, ArticleView } from '../../../../entities/Article';
 import {
     getArticlesPageOrder, getArticlesPageSearch, getArticlesPageSort, getArticlesPageType, getArticlesPageView,
 } from '../../../../pages/ArticlesPage/model/selectors/articlesPageSelectors';
@@ -19,6 +15,8 @@ import { articlesPageActions } from '../../../../pages/ArticlesPage/model/slices
 import { classNames } from '../../../../shared/lib/classNames/classNames';
 import { useAppDispatch } from '../../../../shared/lib/hooks/useAppDispatch/useAppDispatch';
 import clss from './ArticlesPageFilters.module.scss';
+import { ArticleSortSelector } from '../../../../features/ArticleSortSelector';
+import { ArticleViewSelector } from '../../../../features/ArticleViewSelector';
 
 interface ArticlesPageFiltersProps {
    className?: string;
